feat: add Autofix configuration and operations
