+++ packages/control-plane/src/index.ts
@@ -9,6 +9,7 @@ import { createLogger } from "./logger";
 import type { Env } from "./types";
 import type { GitHubAutofixEnvelope } from "@open-inspect/shared";
 import { handleAutofixQueue } from "./autofix/handler";
+import { checkAutofixQueueHealth } from "./autofix/queue-health";
 import { consumeImageBuildFinalizations } from "./image-builds/finalization-consumer";
 import { IMAGE_BUILD_SCHEDULER_CRON, runImageBuildScheduler } from "./image-builds/scheduler";
 import {
@@ -71,6 +72,7 @@ export default {
       logger.warn("Unknown scheduled trigger", { cron: event.cron });
       return;
     }
+    ctx.waitUntil(checkAutofixQueueHealth(env, logger));
     // The tick runs both the recovery sweep (orphaned/timed-out runs) and
     // processes overdue automations.
     // eslint-disable-next-line no-restricted-syntax -- scheduled composition root: construct the scheduler's database dependency
