+++ packages/control-plane/src/types.ts
@@ -16,6 +16,10 @@ export interface Env {
   SLACK_BOT?: Fetcher; // Optional - only if slack-bot is deployed
   LINEAR_BOT?: Fetcher; // Optional - only if linear-bot is deployed
 
+  // GitHub Autofix queue bindings used for read-only metrics.
+  AUTOFIX_QUEUE?: Queue<unknown>;
+  AUTOFIX_DLQ?: Queue<unknown>;
+
   // D1 database
   DB: D1Database;
