+++ packages/web/src/components/session-timeline.test.tsx
@@ -67,6 +67,32 @@ function toolCall(callId: string, tool: string, filePath: string): SandboxEvent 
 }
 
 describe("user message authors", () => {
+  it("presents Autofix provenance and links to the originating review", () => {
+    render(
+      <EventItem
+        event={{
+          ...event("user-2"),
+          origin: {
+            kind: "review",
+            authorType: "bot",
+            feedbackUrl: "https://github.com/acme/widgets/pull/42#pullrequestreview-5678",
+          },
+        }}
+        sessionId="session-1"
+        currentParticipantId="participant-1"
+        participantProfiles={{}}
+        onOpenMedia={() => {}}
+      />
+    );
+
+    expect(screen.getByText("Resumed by PR feedback")).toBeInTheDocument();
+    expect(screen.getByText("Review · Bot")).toBeInTheDocument();
+    expect(screen.getByRole("link", { name: "Open feedback" })).toHaveAttribute(
+      "href",
+      "https://github.com/acme/widgets/pull/42#pullrequestreview-5678"
+    );
+  });
+
   it("uses the canonical profile name and avatar when available", () => {
     render(
       <EventItem
