+++ terraform/environments/production/workers-control-plane.tf
@@ -55,12 +55,26 @@ module "control_plane_worker" {
     }
   ]
 
-  queue_bindings = [
-    {
-      binding_name = "IMAGE_BUILD_FINALIZATION_QUEUE"
-      queue_name   = cloudflare_queue.image_build_finalization.queue_name
-    }
-  ]
+  # These bindings provide read-only queue metrics to the operator health
+  # check. Autofix production remains owned by the GitHub bot.
+  queue_bindings = concat(
+    [
+      {
+        binding_name = "IMAGE_BUILD_FINALIZATION_QUEUE"
+        queue_name   = cloudflare_queue.image_build_finalization.queue_name
+      }
+    ],
+    var.enable_github_bot ? [
+      {
+        binding_name = "AUTOFIX_QUEUE"
+        queue_name   = cloudflare_queue.github_autofix[0].queue_name
+      },
+      {
+        binding_name = "AUTOFIX_DLQ"
+        queue_name   = cloudflare_queue.github_autofix_dlq[0].queue_name
+      }
+    ] : []
+  )
 
   service_bindings = concat(
     var.enable_slack_bot ? [
